Move symex_bmc and symex_coverage to goto-checker
These classes are language-independent.
