+++ src/goto-checker/symex_coverage.cpp
