+++ src/goto-checker/symex_coverage.h
